my-coffee-store/backend/requirements.txt
# my-coffee-store/backend/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4